requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
feedparser==6.0.10
beautifulsoup4==4.12.2
python-dotenv==1.0.0
APScheduler==3.10.4
